adk_helper/pyproject.toml
[project]
name = "adk_helper"
version = "0.1.0"
description = "A helper to build ADK agents."
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "google-adk>=1.3.0",
    "pytest>=8.3.5",
    "pytest-mock>=3.14.1",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["adk_agent"]
